model/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pillow>=10.4.0
numpy>=1.26.0
requests>=2.32.0
torch>=2.1.0
torchvision>=0.16.0